remove duplicated symbolic skeletonization helpers
